Created AbstractPluginComponent
